wip/moved obsidian properties logic to the ofm transformer
