+++ quartz.config.ts
@@ -58,7 +58,6 @@ const config: QuartzConfig = {
       Plugin.TableOfContents(),
       Plugin.CrawlLinks({ markdownLinkResolution: "shortest" }),
       Plugin.Description(),
-      Plugin.PropertyLinks(),
     ],
     filters: [Plugin.RemoveDrafts()],
     emitters: [
