+++ quartz/plugins/transformers/index.ts
@@ -9,4 +9,3 @@ export { OxHugoFlavouredMarkdown } from "./oxhugofm"
 export { SyntaxHighlighting } from "./syntax"
 export { TableOfContents } from "./toc"
 export { HardLineBreaks } from "./linebreaks"
-export { PropertyLinks } from "./propertylinks"
